requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.37.19
botocore==1.37.19
cachetools==5.5.2
catboost==1.2.7
ccxt==4.4.63
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.2.1
cryptography==44.0.1
cycler==0.12.1
databricks-sdk==0.46.0
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
dotenv==0.9.9
dynaconf==3.2.10
evidently==0.6.6
Faker==37.0.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.6.0
litestar==2.15.1
litestar-htmx==0.4.1
lxml==5.3.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
mlflow==2.20.4
mlflow-skinny==2.20.4
msgspec==0.19.0
multidict==6.1.0
multipart==1.2.1
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.29.0
nltk==3.9.1
numpy==1.26.4
nvidia-nccl-cu12==2.26.2
opentelemetry-api==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
patsy==1.0.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
polyfactory==2.19.0
prometheus_client==0.21.1
propcache==0.3.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-click==1.8.8
rsa==4.9
s3transfer==0.11.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
statsmodels==0.14.4
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==1.26.20
uuid6==2024.7.10
uvicorn==0.34.0
uvloop==0.21.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.0.4
websocket==0.2.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==2.1.4
yarl==1.18.3
yfinance==0.2.54
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
